requirements.txt
python-dotenv
llama-cpp-python
vosk
psycopg2-binary
websockets